==> Final_Fig2/STEP5_TandemUTR graph.Rmd <==
---
title: "Step5_Dotplot comparing tandem 3UTR isoforms"
author: "Yeonui+Kwak"
date: "8/8/2020"
output: html_document
---
1.Set a working directory
```{R}
setwd("~/Desktop/Publication_Mar/Final_Fig2/Rcode/")

```

#set up color
```{r}
library(RColorBrewer)
#nstall.packages("raster")
library(raster)
library(tidyverse)
library(dplyr)
library(tidyr)
library(ggplot2)
library(ggbeeswarm)
yk.col = function(n = 5, sat = 0.7, lum = 1) {
	col = c("royalblue4", "deepskyblue3", "turquoise3",
	   "grey62", "goldenrod2", "orange2", "orangered2") %>%
	col2rgb %>%
	rgb2hsv
	return(colorRampPalette(hsv(col[1,], col[2,] * sat,
				    col[3,] * lum))(n)) }
```

1.Select the genes with multiple tandem UTR( 2 and 3 only)

```{R}
final=read.table("custom.pal.withoutnearbyPAS.txt",header=T,stringsAsFactors = F)
colnames(final)
t=final%>%
  filter(meanpal<300)%>%
  group_by(hgnc,ALE)%>%summarise(tandem_count=n())%>%filter(tandem_count%in%c(2,3))%>%
  unite(newid,c(hgnc,ALE),sep=":")
custom3=final%>%
  filter(meanpal<300)%>%
  unite("newid",c(hgnc,ALE),sep=":")%>%
  inner_join(t,by="newid")%>%
  separate(newid,c("hgnc","ALE"),sep=":")%>%arrange(hgnc)

```

#new ratio: readcountofa peak/sum(peaks)
```{R}
custom4=custom3%>%ungroup()%>%group_by(hgnc)%>%mutate(newratio=readcount/sum(readcount))
omit.list=custom4%>%filter(tandem_count==2)%>%select(hgnc,ALE,Tandem,utrlen)%>%spread(Tandem, utrlen)%>%mutate(diff=`2`-`1`)%>%filter(diff<0)%>%unite(hgncale,c("hgnc","ALE"),sep=":")%>%select(hgncale)
custom4=custom4%>%unite(hgncale,c("hgnc","ALE"),sep=":",remove=F)%>%filter(!(hgncale %in% omit.list$hgncale)) 

pdf("Fig2C_CustomTEDseq_TandemUTR_PAL_dotplot.pdf",width=6,height=3)
ggplot(custom4%>%ungroup()%>%filter(tandem_count %in% c(2,3)),aes(x=Tandem,y=meanpal))+geom_quasirandom(method = "smiley",aes(colour=(newratio)),size=0.1) + 
  #geom_beeswarm()+
  scale_colour_gradientn(colours = yk.col(1000),name="RNA expression level")+
  #geom_dotplot(aes(fill=as.factor(ratio),colour=as.factor(ratio)),alpha=1,binaxis = "y", stackdir="center",stackgroups = TRUE, binwidth =1,dotsize =2)+
  facet_wrap(.~tandem_count,ncol=5,scale="free_x")+
  theme_bw()+xlab("")+
  
theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+
xlab((expression("tandem 3´UTR Index (proximal"*symbol('\256')*"distal)")))+ylab("Mean poly(A) tail length (nt)")

  dev.off()
```


#Statisctical difference of PAL between 3UTR proximal vs distal

#For meanpal<=200
```{r}
#library(cowplot)
#
stat=custom3%>%
  filter(meanpal<=200)%>%
  group_by(tandem_count,Tandem)%>%summarise(number=n())%>%
  ungroup()%>%
  mutate(`mean(meanpal)`=custom3%>%filter(meanpal<=200)%>%group_by(tandem_count,Tandem)%>%summarise(t=mean(meanpal))%>%ungroup()%>%select(t)%>%unlist)
custom3%>%group_by(tandem_count)%>%summarise(number=n())
write.table(stat,"tandem3UTR_PALstatlessthan200nt.txt",col.names=T,row.names = F,sep="\t",quote = F)
custom3.1=custom3%>%filter(meanpal<=200&tandem_count==2)
t.test(custom3.1$meanpal[custom3.1$Tandem==1],custom3.1$meanpal[custom3.1$Tandem==2], alternative = "two.sided")
#p-value = 9.397e-08
custom3.1=custom3%>%filter(meanpal<=200&tandem_count==3)
t.test(custom3.1$meanpal[custom3.1$Tandem==1|custom3.1$Tandem==2],custom3.1$meanpal[custom3.1$Tandem==3], alternative = "two.sided")
#p-value = 5.361e-05

```


Fig2E
```{R}
over200.mean=custom4%>%mutate(group=ifelse(meanpal>=200,"extremely long","≤200 nt"))%>%
  group_by(tandem_count,Tandem,group)%>%
  summarise(mean=mean(newratio),sd=sd(newratio),n=n())

df.summary2 <- custom4%>%ungroup()%>%mutate(PALgroup=ifelse(meanpal>=200,"extremely long","≤200 nt"))%>%select(tandem_count,Tandem,PALgroup,newratio)%>%
  group_by(tandem_count,Tandem,PALgroup)%>%
  summarise(
    sd = sd(newratio),
    RNA = mean(newratio),
    n=n()
  )
pdf("Fig2E.pdf",width=6,height=3 )
ggplot(df.summary2%>%ungroup()%>%mutate(Tandem=as.factor(Tandem)), aes(Tandem, RNA)) +
  geom_col(aes(fill = PALgroup), position = position_dodge(0.8), width = 0.7)+
  geom_errorbar(
    aes(ymin = RNA, ymax = RNA+sd, group = PALgroup),
    width = 0.2, position = position_dodge(0.8)
    )+facet_grid(.~tandem_count,scale="free",space="free_x")+theme_bw()+scale_fill_manual(name="Poly(A) tail length",values=yk.col(9)[c(7,3)])+ylim(c(0,1.1))+xlab((expression("tandem 3´UTRs (proximal"*symbol('\256')*"distal)")))+ylab("Mean of normalized RNA abundance")
dev.off()
df.summary3=custom4%>%ungroup()%>%filter(tandem_count==2)%>%mutate(Tandem=as.factor(Tandem))
pdf("Fig2E_1.pdf",width=4,height=3 )
ggplot(df.summary3, aes(Tandem, RNA)) +
  geom_col(aes(fill = PALgroup), position = position_dodge(0.8), width = 0.7)+
  geom_errorbar(
    aes(ymin = RNA, ymax = RNA+sd, group = PALgroup),
    width = 0.2, position = position_dodge(0.8)
    )+theme_bw()+scale_fill_manual(name="Poly(A) tail length",values=yk.col(9)[c(7,3)])+ylim(c(0,1.1))+xlab((expression("tandem 3´UTRs (proximal"*symbol('\256')*"distal)")))+ylab("Mean of normalized RNA abundance")
dev.off()

write.table(df.summary2,"Summary_GT200count.tandemUTRgroup.txt",col.names = T,row.names = F,sep="\t",quote=F)
```

Fig2D
```{r}

over200=custom4%>%group_by(tandem_count,Tandem)%>%filter(meanpal>=200)%>%summarise(number=n())

over200 
pdf("Fic2D_Number of tandem UTR with PAL greater than 200.pdf",width=2,height=3)
ggplot(over200%>%filter(tandem_count==2),aes(x=as.factor(Tandem),y=number,fill=as.factor(Tandem)))+
  #facet_wrap(.~tandem_count,ncol=4,scale="free_x")+
  geom_bar(stat="identity")+scale_fill_manual(values=yk.col(9)[c(9,7,5,4)])+theme_bw()+xlab((expression("tandem 3´UTR index")))+ylab("Number of tandem 3'UTRs \n with PAL ≥200 nt")+theme(legend.position = "none")
  #+scale_x_discrete(limits=c("1", "2","3","4"))
dev.off()



pdf("FicS2D_Number of tandem UTR with PAL greater than 200.pdf",width=4,height=3)
ggplot(over200%>%filter(tandem_count %in% c(2,3)),aes(x=as.factor(Tandem),y=number,fill=as.factor(Tandem)))+facet_wrap(.~tandem_count,ncol=4,scale="free_x")+ geom_bar(stat="identity")+scale_fill_manual(values=yk.col(9)[c(9,7,5,4)])+theme_bw()+xlab((expression("tandem 3´UTR index")))+ylab("Number of tandem 3'UTRs \n with PAL ≥200 nt")+theme(legend.position = "none")
#+scale_x_discrete(limits=c("1", "2","3","4"))
dev.off()
```


n=2
```{R}
tandem.pal.len=custom4%>%filter(meanpal<=200)%>%
  filter(tandem_count==2)%>%select(hgnc,ALE,Tandem,meanpal)%>%
  #mutate_at(scale, .vars = vars(meanpal))%>%
  spread(Tandem, meanpal)%>%
  mutate(diff=`2`-`1`)%>%mutate(type="Poly(A) tail length")

t.test(tandem.pal.len$diff) #p-value = 1.302e-07

tandem.rna=custom4%>%filter(meanpal<=200)%>%
  filter(tandem_count==2)%>%
  #group_by(hgnc,ALE)%>%
  #mutate(newratio=mean.rc/sum(mean.rc))%>%
  select(hgnc,ALE,Tandem,newratio)%>%
  #mutate_at(scale, .vars = vars(readcount))%>%
  spread(Tandem, newratio)%>%mutate(diff=`2`-`1`)%>%mutate(type="RNA abundance")
t.test(tandem.rna$diff) #p-value = 3.109e-06

tandem.utrlen=custom4%>%filter(meanpal<=200)%>%
  filter(tandem_count==2)%>%select(hgnc,ALE,Tandem,utrlen)%>%
  #mutate_at(scale, .vars = vars(utrlen))%>%
  spread(Tandem, utrlen)%>%
  mutate(diff=`2`-`1`)%>%mutate(type="3´UTR length")

t.test(tandem.utrlen$diff)
#tandem.summary=bind_rows(tandem.rna,tandem.pal.len,tandem.utrlen)%>%na.omit()
#tandem.summary%>%filter(type=="Poly(A) tail length")


p1<-ggplot(tandem.utrlen,aes(x=type,y=diff,fill=type))+
  geom_boxplot(notch = T)+
  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(9)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+#ylim(c(-2,2))+
 ylab(expression(Delta*"3´UTR length (distal- proximal)"))+xlab("")+ylim(c(0,4000))


p2<-ggplot(tandem.pal.len,aes(x=type,y=diff,fill=type))+
  geom_boxplot(notch = T)+
  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(7)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+#ylim(c(-2,2))+
 ylab(expression(Delta*"PAL (distal- proximal)"))


p3<-ggplot(tandem.rna,aes(x=type,y=diff,fill=type))+
  geom_boxplot(notch = T)+

  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(3)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+#ylim(c(-2,2))+
 ylab(expression(Delta*"RNA (distal- proximal)"))
pdf("Fig2F_tandem.summary.pdf",width=6,height=4)
grid.arrange(
  p1,
  p2,
  p3,
  nrow = 1,
  top= "Pairwise comparison of tandem 3´UTR pairs"
  )

dev.off()
```

```{R}
tandem.pal.len=custom4%>%filter(meanpal<=200)%>%
  filter(tandem_count==3)%>%select(hgnc,ALE,Tandem,meanpal)%>%
  #mutate_at(scale, .vars = vars(meanpal))%>%
  spread(Tandem, meanpal)%>%
  mutate(diff=`3`-`1`)%>%mutate(type="Poly(A) tail length")

t.test(tandem.pal.len$diff) #p-value = 1.302e-07

tandem.rna=custom4%>%filter(meanpal<=200)%>%
  filter(tandem_count==3)%>%
  #group_by(hgnc,ALE)%>%
  #mutate(newratio=mean.rc/sum(mean.rc))%>%
  select(hgnc,ALE,Tandem,newratio)%>%
  #mutate_at(scale, .vars = vars(readcount))%>%
  spread(Tandem, newratio)%>%mutate(diff=`3`-`1`)%>%mutate(type="RNA abundance")
t.test(tandem.rna$diff) #p-value = 3.109e-06

tandem.utrlen=custom4%>%filter(meanpal<=200)%>%
  filter(tandem_count==3)%>%select(hgnc,ALE,Tandem,utrlen)%>%
  #mutate_at(scale, .vars = vars(utrlen))%>%
  spread(Tandem, utrlen)%>%
  mutate(diff=`3`-`1`)%>%mutate(type="3´UTR length")

t.test(tandem.utrlen$diff)
#tandem.summary=bind_rows(tandem.rna,tandem.pal.len,tandem.utrlen)%>%na.omit()
#tandem.summary%>%filter(type=="Poly(A) tail length")


p4<-ggplot(tandem.utrlen,aes(x=type,y=diff,fill=type))+
  geom_boxplot(notch = T)+
  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(9)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+#ylim(c(-2,2))+
 ylab(expression(Delta*"3´UTR length (distal- proximal)"))+xlab("")+ylim(c(0,4000))


p5<-ggplot(tandem.pal.len,aes(x=type,y=diff,fill=type))+
  geom_boxplot(notch = T)+
  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(7)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+#ylim(c(-2,2))+
 ylab(expression(Delta*"PAL (distal- proximal)"))


p6<-ggplot(tandem.rna,aes(x=type,y=diff,fill=type))+
  geom_boxplot(notch = T)+

  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(3)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+#ylim(c(-2,2))+
 ylab(expression(Delta*"RNA (distal- proximal)"))
pdf("tandem.summary2.pdf",width=5,height=6)
grid.arrange(
  p1,
  p2,
  p3,
  p4,
  p5,
  p6,
  nrow = 2,
  top= "Pairwise comparison of tandem 3´UTR pairs"
  )

dev.off()
```







#useless
```{R}
pdf("CustomTEDseq_TandemUTR_Abundance vs.medianPAL.pdf",width=10,height=3)
p<-ggplot(custom3,aes(x=new.tandem,y=meanpal,fill=as.factor(new.tandem)))+geom_boxplot(notch = T)+
facet_wrap(.~tandem_count,ncol=5,scale="free_x")+
  #stat_summary(fun.data = stat_box_data, geom = "text", hjust = 0.5,vjust = 0.9x) + 
  #geom_dotplot(alpha=1,binaxis = "y", stackgroups = TRUE, binwidth =1,dotsize =2, stackdir="center")
  theme_bw()+theme(legend.position = "none")+xlab("")+
scale_fill_manual(values=yk.col(9)[c(9,7,5:1)])+theme(axis.title.x = element_text(size = 10, # also adjust text size if needed
                                    margin = margin(t = 10, r = 0, b = 0, l = 0,
                                                    unit = "mm")))+
  xlab((expression("tandem 3´UTRs (lowly expressed"*symbol('\256')*"highly expressed)")))+ylab("Mean poly(A) tail length (nt)")
ggdraw(p)#+
  #draw_label(line1, x = 0.55, y = 0.075) + # use relative coordinates for positioning
  #draw_label(line2, x = 0.55, y = 0.025)
  dev.off()
```
